services/muse-compute-engine/requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic
websockets

# Math/Science
sympy
numpy
scipy
pandas

# Database
asyncpg
psycopg2-binary

# Physics (unit-aware math)
pint

# Chemistry (SMILES/molecular)
rdkit-pypi

# NLP/Text
networkx
nltk
spacy
biopython

# Visualization
matplotlib
